Get input data

Get input data

Usage

get_input(
  x,
  types = get_types(x),
  samples = get_samples(x),
  clusters = get_cluster_labels(x),
  matrix = FALSE,
  reconstructed = FALSE,
  add_groups = FALSE,
  by_sigs = FALSE
)

Arguments

x

bascule object.

types

List of variant types to subset

samples

List of samples to subset

clusters

List of clusters to subset

matrix

Logical. If `TRUE`, the output will be in wide format.

reconstructed

Logical. If `TRUE`, the reconstructed counts, i.e., computed as alpha x beta, will be returned.

add_groups

Logical. If `TRUE` and clustering has been performed, each sample will be annotated with the assigned group.

by_sigs

Logical. If `TRUE`, the number of mutations generated by each signature will be returned.

Value

list of long or wide-formatted mutations counts.